Move code to middleware.go
